Move rate estimate from 64-bits to 32-bits, saving some bytes.

+++ src/rust/lqos_bus/src/ip_stats.rs
@@ -149,7 +149,7 @@ pub struct FlowbeeData {
   /// Packets transmitted
   pub packets_sent: [u64; 2],
   /// Rate estimate
-  pub rate_estimate_bps: [u64; 2],
+  pub rate_estimate_bps: [u32; 2],
   /// TCP Retransmission count (also counts duplicates)
   pub tcp_retransmits: [u32; 2],
   /// Most recent RTT
